Opciones de la funcion heuristica correspondiente y algunas pruebas

+++ include/environment.hpp
@@ -43,17 +43,18 @@ class Environment {
 
   void move_car(cardinal x, int steps);
 
-  int lineal_d(void);  // Funcion heuristica lineal por defecto (coche - meta)
-  int manhattan_d(
+  double lineal_d(
+      void);  // Funcion heuristica lineal por defecto (coche - meta)
+  double manhattan_d(
       void);  // Funcion heuristica manhattan por defecto (coche - meta)
 
-  int lineal_d(const Slot& begin);     // Funcion heuristica lineal por defecto
+  double lineal_d(const Slot& begin);  // Funcion heuristica lineal por defecto
                                        // (Slot - meta)
-  int manhattan_d(const Slot& begin);  // Funcion heuristica manhattan por
-                                       // defecto (Slot - meta)
-  int lineal_d(const Slot& begin,
-               const Slot& end);  // Funcion heuristica lineal (Slot -Slot)
-  int manhattan_d(
+  double manhattan_d(const Slot& begin);  // Funcion heuristica manhattan por
+                                          // defecto (Slot - meta)
+  double lineal_d(const Slot& begin,
+                  const Slot& end);  // Funcion heuristica lineal (Slot -Slot)
+  double manhattan_d(
       const Slot& begin,
       const Slot& end);  // Funcion heuristica manhattan (Slot -Slot)
